% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Coordinates.Validator.R
\name{Coordinates.Validator}
\alias{Coordinates.Validator}
\title{Coordinates Validator}
\usage{
Coordinates.Validator()
}
\value{
A list of validator functions to validate different aspects of coordinates. :

\describe{
\item{\code{Coordinates}}{Validate if coordinates exists and has x and y attributes.}
\item{\code{exists}}{Validate if coordinates exists.}
\item{\code{has.x}}{Validates if coordinates has x attribute.}
\item{\code{has.y}}{Validates if coordinates has y attribute.}
}
}
\description{
This component contains all needed validators to check if coordinates are .valid.factor(
}
