% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{draw_christmas_tree}
\alias{draw_christmas_tree}
\title{Draw a Christmas tree}
\usage{
draw_christmas_tree(n)
}
\arguments{
\item{n}{The number of rows in the Christmas tree.}
}
\value{
None
}
\description{
This function draws a Christmas tree with the specified number of rows.
}
\examples{
draw_christmas_tree(5)

}
