% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{draw_right_triangle}
\alias{draw_right_triangle}
\title{Draw a right triangle}
\usage{
draw_right_triangle(n)
}
\arguments{
\item{n}{The number of rows in the right triangle.}
}
\value{
None
}
\description{
This function draws a right triangle with the specified number of rows.
}
\examples{
draw_right_triangle(5)

}
