% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{draw_square}
\alias{draw_square}
\title{Draw a square}
\usage{
draw_square(n)
}
\arguments{
\item{n}{The number of rows and columns in the square.}
}
\value{
None
}
\description{
This function draws a square with the specified number of rows and columns.
}
\examples{
draw_square(5)

}
