\name{gels}
\alias{gels}
\title{Electrophoresis gel data}
\description{Electrophoresis gel data. 10 invariant spots have been picked out by an expert
on two electrophoretic gels. 
}
\usage{data(gels)}
\format{
An array of dimension 10 x 2 x 2
}
\source{
Dryden, I. L. and Walker, G. (1999). Highly resistant regression and object matching. Biometrics, 55,
820-825.
}
\references{
Data from Chris Glasbey (BioSS)
}
\examples{
data(gels)
plotshapes(gels)
}
\keyword{datasets}

