% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.Generator.R
\name{Generator}
\alias{Generator}
\title{Shape Generator}
\usage{
Generator()
}
\value{
A list of validator functions, which includes:
\describe{
\item{\code{Line.horizontal}}{Use length to generate x y coordinates to plot a horizontal line.}
\item{\code{Line.vertical}}{Use length to generate x y coordinates to plot a vertical line.}
\item{\code{Rectangle}}{Use width and height to generate x y coordinates to plot a rectangle.}
\item{\code{Triangle}}{Use base and height to generate x y coordinates to plot a triangle.}
\item{\code{Circle}}{Use radius to generate x y coordinates to plot a circle.}
\item{\code{Trapezoid}}{Use bottom, top and height to generate x y coordinates to plot a trapezoid.}
\item{\code{Segment}}{Use radius to generate x y coordinates to plot a segment.}
}
}
\description{
This Generator Component returns a collection of geometric shape generator functions.
Each generator function is designed to create x, y coordinate pairs for specific geometric shapes
based on given specifications. It uses validators to ensure that the specifications meet
the requirements.
}
