\name{gorf.dat}
\alias{gorf.dat}
\title{Female gorilla data}
\description{Female gorilla skull data. 8 landmarks 
in 2 dimensions, 30 individuals
}
\usage{data(gorf.dat)}
\format{
An array of dimension 8 x 2 x 30
}
\source{
Dryden, I.L. and Mardia, K.V. (1998). 
Statistical Shape Analysis, Wiley, Chichester. p317 

O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
http://www.maths.nott.ac.uk/personal/ild/bookdata/gorf.dat

Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(gorf.dat)
plotshapes(gorf.dat)
}
\keyword{datasets}

