% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Offset.Validator.R
\name{Offset.Validator}
\alias{Offset.Validator}
\title{Offset Validator}
\usage{
Offset.Validator()
}
\value{
A list of validator functions to validate different aspects of offset:

\describe{
\item{\code{Offset}}{Validate if offset exists and has x and y attributes.}
\item{\code{exists}}{Validate if offset exists.}
\item{\code{has.x}}{Validates if offset has x attribute.}
\item{\code{has.y}}{Validates if offset has y attribute.}
}
}
\description{
This component contains all needed validators to check if an offset is.
}
