\name{panf.dat}
\alias{panf.dat}
\title{Female chimpanzee data}
\description{Female chimpanzee skull data. 8 landmarks
in 2 dimensions, 26 individuals
}
\usage{data(panf.dat)}
\format{
An array of dimension 8 x 2 x 26
}
\source{
O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(panf.dat)
plotshapes(panf.dat)
}
\keyword{datasets}

