\name{panm.dat}
\alias{panm.dat}
\title{Male chimpanzee data}
\description{Male chimpanzee skull data. 8 landmarks
in 2 dimensions, 28 individuals
}
\usage{data(panm.dat)}
\format{
An array of dimension 8 x 2 x 28
}
\source{
O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(panm.dat)
plotshapes(panm.dat)
}
\keyword{datasets}

