\name{pongof.dat}
\alias{pongof.dat}
\title{Female orang utan data}
\description{Female orang utan skull data. 8 landmarks
in 2 dimensions, 30 individuals
}
\usage{data(pongof.dat)}
\format{
An array of dimension 8 x 2 x 30
}
\source{
O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(pongof.dat)
plotshapes(pongof.dat)
}
\keyword{datasets}

