\name{pongom.dat}
\alias{pongom.dat}
\title{Male orang utan data}
\description{Male orang utan skull data. 8 landmarks
in 2 dimensions, 30 individuals
}
\usage{data(pongom.dat)}
\format{
An array of dimension 8 x 2 x 30
}
\source{
O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(pongom.dat)
plotshapes(pongom.dat)
}
\keyword{datasets}

