% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shape.R
\name{print_diamond}
\alias{print_diamond}
\title{Print a diamond}
\usage{
print_diamond(n)
}
\arguments{
\item{n}{The number of rows in the diamond}
}
\value{
None
}
\description{
This function prints a diamond shape with the specified number of rows
}
\examples{
print_diamond(5)

}
