% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Rectangle.Coordinates.Generator.R
\name{Rectangle.Coordinates.Generator}
\alias{Rectangle.Coordinates.Generator}
\title{Rectangle.Coordinates.Generator'}
\value{
a data.frame with the coordinates of a rectangle.
}
\description{
Generates a data.frame with the coordinates of a rectangle.
}
