% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Shape.Generator.R
\name{Shape.Generator}
\alias{Shape.Generator}
\title{Shape.Generator}
\description{
Generates coordinates for primary shapes:
\itemize{
\item Rectangle
\item Triangle
\item Circle
\item Trapezoid
\item Segment
}
}
