\name{gorm.dat}
\alias{gorm.dat}
\title{Male gorilla data}
\description{Male gorilla skull data. 8 landmarks 
in 2 dimensions, 29 individuals
}
\usage{data(gorm.dat)}
\format{
An array of dimension 8 x 2 x 29
}
\source{Dryden, I.L. and Mardia, K.V. (2016). Statistical 
Shape Analysis, with Applications in R (Second Edition). Wiley, Chichester. Chapter 1.

O'Higgins, P. and Dryden, I. L. (1993). Sexual dimorphism in
hominoids: further studies of craniofacial shape differences in
Pan, Gorilla, Pongo, Journal of Human Evolution, 24, 183-205.
}
\references{
http://www.maths.nott.ac.uk/personal/ild/bookdata/gorm.dat

Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(gorm.dat)
plotshapes(gorm.dat)
}
\keyword{datasets}

