\name{macf.dat}
\alias{macf.dat}
\title{Female macaque data}
\description{Female macaque skull data. 7 landmarks 
in 3 dimensions, 9 individuals
}
\usage{data(macf.dat)}
\format{
An array of dimension 7 x 3 x 9
}
\source{
Dryden, I.L. and Mardia, K.V. (2016). Statistical 
Shape Analysis, with Applications in R (Second Edition). Wiley, Chichester. Chapter 1.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(macf.dat)
plotshapes(macf.dat)
}
\keyword{datasets}

