\name{macm.dat}
\alias{macm.dat}
\title{Male macaque data}
\description{Male macaque skull data. 7 landmarks 
in 3 dimensions, 9 individuals
}
\usage{data(macm.dat)}
\format{
An array of dimension 7 x 3 x 9
}
\source{
Dryden, I.L. and Mardia, K.V. (2016). Statistical 
Shape Analysis, with Applications in R (Second Edition). Wiley, Chichester. Chapter 1.
}
\references{
Data from Paul O'Higgins (Hull-York Medical School)
}
\examples{
data(macm.dat)
plotshapes(macm.dat)
}
\keyword{datasets}

