\name{qcet2.dat}
\alias{qcet2.dat}
\title{Control T2 mouse vertabrae data}
\description{T2 mouse vertebrae data - control group. 6 landmarks 
in 2 dimensions, 30 individuals
}
\usage{data(qcet2.dat)}
\format{
An array of dimension 6 x 2 x 30
}
\source{
Dryden, I.L. and Mardia, K.V. (2016). Statistical 
Shape Analysis, with Applications in R (Second Edition). Wiley, Chichester. Chapter 1.
}
\references{
http://www.maths.nott.ac.uk/personal/ild/bookdata/qcet2.dat

Data from Paul O'Higgins (Hull-York Medical School) and David Johnson (Leeds)
}
\examples{
data(qcet2.dat)
plotshapes(qcet2.dat)
}
\keyword{datasets}

