\name{qlet2.dat}
\alias{qlet2.dat}
\title{Large T2 mouse vertabrae data}
\description{T2 mouse vertebrae data - large group. 6 landmarks 
in 2 dimensions, 23 individuals
}
\usage{data(qlet2.dat)}
\format{
An array of dimension 6 x 2 x 23
}
\source{
Dryden, I.L. and Mardia, K.V. (2016). Statistical 
Shape Analysis, with Applications in R (Second Edition). Wiley, Chichester. Chapter 1.
}
\references{
http://www.maths.nott.ac.uk/personal/ild/bookdata/qlet2.dat

Data from Paul O'Higgins (Hull-York Medical School) and David Johnson (Leeds)
}
\examples{
data(qlet2.dat)
plotshapes(qlet2.dat)
}
\keyword{datasets}

