\name{shapes-internal}
\alias{distCholesky}
\alias{distEuclidean}
\alias{distLogEuclidean}
\alias{distPowerEuclidean}
\alias{distProcrustesFull}
\alias{distProcrustesSizeShape}
\alias{distRiemPennec}
\alias{Enorm}
\alias{estChol}
\alias{estShapes}
\alias{estEuclid}
\alias{estLogEuclid}
\alias{estPowerEuclid}
\alias{estLogRiem2}
\alias{distRiemannianLe} 
\alias{estCholesky}
\alias{estRiemLe}
\alias{estShape}
\alias{estSS}
\alias{Hessian2}
\alias{iglogl}
\alias{Lambdamin}
\alias{nsa}
\alias{protein}
\alias{James}
\alias{procWGPA1}
\alias{rotate}
\alias{bootstraptest}
\alias{testmeanshapes.old}
\alias{permutationtest2}
\alias{Goodall}
\alias{Hotelling}
\alias{abind}
\alias{tpsgrid.old}
\alias{shaperw}
\alias{BoxM} 
\alias{Goodall2D}
\alias{Goodalltest} 
\alias{Hotelling2D}
\alias{Hotelling2Djames}
\alias{Hotellingtest}
\alias{MGM}
\alias{I2mat} 
\alias{TPSgrid}
\alias{V} 
\alias{Vinv} 
\alias{Vmat} 
\alias{add} 
\alias{as.3d}
\alias{banner1} 
\alias{banner4} 
\alias{bendingenergy} 
\alias{bgpa} 
\alias{bookstein.shpv} 
\alias{bookstein.shpv.complex}
\alias{braincon.dat}
\alias{brainscz.dat}
\alias{cbevec}
\alias{cbevectors}
\alias{centroid.size}
\alias{centroid.size.complex} 
\alias{centroid.size.mD} 
\alias{close1} 
\alias{cnt3} 
\alias{complextoreal} 
\alias{defh} 
\alias{defplotsize2}
\alias{defplotsize3} 
\alias{del} 
\alias{dif} 
\alias{dif.old} 
\alias{dis}
\alias{distcov}
\alias{estcov}
\alias{fJ}
\alias{fcel} 
\alias{fcnt} 
\alias{fgpa} 
\alias{fgpa.rot} 
\alias{fgpa.singleiteration} 
\alias{fopa} 
\alias{fort.ROTATEANDREFLECT} 
\alias{fort.ROTATION} 
\alias{fort} 
\alias{fos} 
\alias{fos.REFLECT}
\alias{ftrsq} 
\alias{full.procdist} 
\alias{genpower} 
\alias{goodall2d} 
\alias{goodalltest} 
\alias{graf} 
\alias{hotelling2d} 
\alias{hotellingtest} 
\alias{isodens} 
\alias{isologdens}
\alias{isomle} 
\alias{isotropy.test} 
\alias{kendall.shpv} 
\alias{linegrid} 
\alias{loglikeiso} 
\alias{loglikeiso2} 
\alias{loneFone} 
\alias{mahpreshapedist} 
\alias{makearray}
\alias{movie} 
\alias{msh} 
\alias{norm} 
\alias{objfuniso} 
\alias{oneFone}
\alias{partial.procdist}
\alias{partialwarpgrids}
\alias{partialwarps}
\alias{permutationtest}
\alias{plot2rwscores}
\alias{plot3Ddata}
\alias{plot3Ddata.static}
\alias{plot3Dmean} 
\alias{plot3Dpca}
\alias{plotPDM} 
\alias{plotPDM2} 
\alias{plotPDM3} 
\alias{plotPDMbook} 
\alias{plotPDMnoaxis} 
\alias{plotPDMnoaxis3}
\alias{plotpairscores} 
\alias{plotpca}
\alias{plotpca3d}
\alias{plotprinwarp} 
\alias{plotproc} 
\alias{plotrelwarp}
\alias{plotshapes3d}
\alias{plotshapestime3d}
\alias{pointsPDMnoaxis3} 
\alias{prcomp1} 
\alias{preshape} 
\alias{preshape.mD} 
\alias{preshape.mat} 
\alias{preshapetoicon} 
\alias{prinwscoregrids} 
\alias{procdistreflect} 
\alias{procrustes2d}
\alias{procrustesGPA} 
\alias{procrustesGPA.rot} 
\alias{procrustesgpa} 
\alias{project} 
\alias{read.array} 
\alias{read.in} 
\alias{realtocomplex} 
\alias{reassqpr} 
\alias{relwarps} 
\alias{rgpa}
\alias{riemdist.complex} 
\alias{riemdist.mD} 
\alias{rotateaxes}
\alias{schizo.dat}
\alias{sgpa}
\alias{sh}
\alias{sigma}
\alias{sim1}
\alias{st}
\alias{tanfigure}
\alias{tanfigurefull}
\alias{tanpreshape}
\alias{tantofigure}
\alias{tantofigurefull}
\alias{tantopreshape}
\alias{testshapes}
\alias{vec1}
\title{Internal function(s)}
\description{Internal function(s)}
\keyword{internal}

