\name{shapes-internal}
\alias{projectPNS}
\alias{sphere1.f}
\alias{rgl.sphgrid1}
\alias{sph2car1}
\alias{pc2sphere2} 
\alias{pcscore2sphere2}
\alias{preshape2shape}
\alias{tangentcoords.partial.inv}
\alias{multiply_by_transpose_of_helmert_implicitly} 
\alias{multiply_by_helmert_implicitly_3d}
\alias{uji3_centroid.size}
\alias{uji2_centroid.size}
\alias{multiply_by_transpose_of_helmert_explicitly} 
\alias{multiply_by_transpose_of_helmert} 
\alias{multiply_by_helmert_implicitly}
\alias{multiply_by_helmert_explicitly}
\alias{multiply_by_helmert}
\alias{uji_preshape}
\alias{uji_centroid.size}
\alias{uji_defh}
\alias{uji_Enorm}
\alias{uji_distProcrustesFull}
\alias{uji_distProcrustesSizeShape}
\alias{uji_distCholesky}
\alias{uji_estSS}
\alias{uji_estShape}
\alias{uji_centroid.size.complex}
\alias{uji_centroid.size.mD}
\alias{uji_preshape.mD}
\alias{uji_preshape.mat}
\alias{uji_tanfigure}
\alias{uji_tanfigurefull}
\alias{uji_kendall.shpv}
\alias{ild_preshape}
\alias{ild_centroid.size}
\alias{ild_defh}
\alias{ild_Enorm}
\alias{ild_distProcrustesFull}
\alias{ild_distProcrustesSizeShape}
\alias{ild_distCholesky}
\alias{ild_estSS}
\alias{ild_estShape}
\alias{ild_centroid.size.complex}
\alias{ild_centroid.size.mD}
\alias{ild_preshape.mD}
\alias{ild_preshape.mat}
\alias{ild_tanfigure}
\alias{ild_tanfigurefull}
\alias{ild_kendall.shpv}
\alias{ild_preshapetoicon}
  \alias{Enormalize} \alias{ExpNPd} \alias{LRTpval} \alias{LogNPd} \alias{PNSe2s} \alias{PNSs2e} \alias{Plot3D}
  \alias{Procrustes.dist.full} \alias{UNIFORMdirections} \alias{col2RGB} \alias{flipud0}
  \alias{geodmeanS1} \alias{get.data.subsphere} \alias{get.prinarc}
  \alias{get.prinarc.subsphere} \alias{get.prinarc.value} \alias{getSubSphere} \alias{mod}
  \alias{objfn} \alias{pc2sphere} \alias{pcscore2sphere} \alias{plotshapes3d.pns}  \alias{pns.pc}
   \alias{project.subsphere} \alias{randvonMisesFisherm} \alias{repmat} \alias{rot.mat}
  \alias{rotMat} \alias{shape.pcscores} \alias{shape.pcscores.partial} \alias{sphere.jac}
  \alias{sphere.obj} \alias{sphere.res} \alias{sphere2pcscore} \alias{sphereFit}
  \alias{tangent.coords.partial} \alias{tr} \alias{trans.subsphere} \alias{vMFtest}
\alias{ped}
\alias{pedreg}
\alias{sooty.dat}
\alias{MDSshape}
\alias{distCholesky}
\alias{distEuclidean}
\alias{distLogEuclidean}
\alias{distPowerEuclidean}
\alias{distProcrustesFull}
\alias{distProcrustesSizeShape}
\alias{rootmat}
\alias{distRiemPennec}
\alias{Enorm}
\alias{estChol}
\alias{estShapes}
\alias{estEuclid}
\alias{estLogEuclid}
\alias{estPowerEuclid}
\alias{estLogRiem2}
\alias{distRiemannianLe} 
\alias{estCholesky}
\alias{estRiemLe}
\alias{estShape}
\alias{estSS}
\alias{Hessian2}
\alias{iglogl}
\alias{Lambdamin}
\alias{nsa}
\alias{protein}
\alias{James}
\alias{select}
\alias{procWGPA1}
\alias{rotatexyz}
\alias{objfun}
\alias{objfun4}
\alias{bootstraptest}
\alias{testmeanshapes.old}
\alias{permutationtest2}
\alias{Goodall}
\alias{Hotelling}
\alias{abind}
\alias{tpsgrid.old}
\alias{shaperw}
\alias{BoxM} 
\alias{Goodall2D}
\alias{Goodalltest} 
\alias{Hotelling2D}
\alias{Hotelling2Djames}
\alias{Hotellingtest}
\alias{MGM}
\alias{I2mat} 
\alias{TPSgrid}
\alias{V} 
\alias{Vinv} 
\alias{Vmat} 
\alias{add} 
\alias{as.3d}
\alias{banner1} 
\alias{banner4} 
\alias{bendingenergy} 
\alias{bgpa} 
\alias{bookstein.shpv} 
\alias{bookstein.shpv.complex}
\alias{braincon.dat}
\alias{brainscz.dat}
\alias{cbevec}
\alias{cbevectors}
\alias{centroid.size.complex} 
\alias{centroid.size.mD} 
\alias{close1} 
\alias{cnt3} 
\alias{complextoreal} 
\alias{defh} 
\alias{defplotsize2}
\alias{defplotsize3} 
\alias{del} 
\alias{dif} 
\alias{dif.old} 
\alias{dis}
\alias{fJ}
\alias{fcel} 
\alias{fcnt} 
\alias{fgpa} 
\alias{fgpa.rot} 
\alias{fgpa.singleiteration} 
\alias{fopa} 
\alias{fort.ROTATEANDREFLECT} 
\alias{fort.ROTATION} 
\alias{fort} 
\alias{fos} 
\alias{fos.REFLECT}
\alias{ftrsq} 
\alias{full.procdist} 
\alias{genpower} 
\alias{goodall2d} 
\alias{goodalltest} 
\alias{graf} 
\alias{hotelling2d} 
\alias{hotellingtest} 
\alias{isodens} 
\alias{isologdens}
\alias{isomle} 
\alias{isotropy.test} 
\alias{kendall.shpv} 
\alias{linegrid} 
\alias{loglikeiso} 
\alias{loglikeiso2} 
\alias{loneFone} 
\alias{mahpreshapedist} 
\alias{makearray}
\alias{movie} 
\alias{msh} 
\alias{norm} 
\alias{objfuniso} 
\alias{oneFone}
\alias{partial.procdist}
\alias{partialwarpgrids}
\alias{partialwarps}
\alias{permutationtest}
\alias{plot2rwscores}
\alias{plot3Ddata}
\alias{plot3Ddata.static}
\alias{plot3Dmean} 
\alias{plot3Dpca}
\alias{plotPDM} 
\alias{plotPDM2} 
\alias{plotPDM3} 
\alias{plotPDMbook} 
\alias{plotPDMnoaxis} 
\alias{plotPDMnoaxis3}
\alias{plotpairscores} 
\alias{plotpca}
\alias{plotpca3d}
\alias{plotprinwarp} 
\alias{plotproc} 
\alias{plotrelwarp}
\alias{plotshapes3d}
\alias{plotshapestime3d}
\alias{pointsPDMnoaxis3} 
\alias{prcomp1} 
\alias{preshape} 
\alias{preshape.mD} 
\alias{preshape.mat} 
\alias{preshapetoicon} 
\alias{prinwscoregrids} 
\alias{procdistreflect} 
\alias{procrustes2d}
\alias{procrustesGPA} 
\alias{procrustesGPA.rot} 
\alias{procrustesgpa} 
\alias{project} 
\alias{read.array} 
\alias{read.in} 
\alias{realtocomplex} 
\alias{reassqpr} 
\alias{relwarps} 
\alias{rgpa}
\alias{riemdist.complex} 
\alias{riemdist.mD} 
\alias{rotateaxes}
\alias{schizo.dat}
\alias{sgpa}
\alias{sh}
\alias{sigmacov}
\alias{sim1}
\alias{st}
\alias{tanfigure}
\alias{tanfigurefull}
\alias{tanpreshape}
\alias{testshapes}
\alias{vec1}
\title{Internal function(s)}
\description{Internal function(s)}
\keyword{internal}

